+++ kit/dapp/src/components/blocks/auth/wallet-security-client.tsx
@@ -1,21 +1,13 @@
 "use client";
 
+import { PincodeDialog } from "@/components/blocks/auth/pincode-dialog";
+import { useState, type ReactNode } from "react";
+import { SetupTwoFactorDialog } from "./setup-two-factor-dialog";
 import {
-  PincodeForm,
-  type PincodeFormValues,
-} from "@/components/blocks/auth/pincode-form";
-import {
-  Dialog,
-  DialogContent,
-  DialogDescription,
-  DialogHeader,
-  DialogTitle,
-} from "@/components/ui/dialog";
-import { useRouter } from "@/i18n/routing";
-import { setPincode } from "@/lib/mutations/user/set-pincode-action";
-import { useTranslations } from "next-intl";
-import { type ReactNode, useState } from "react";
-import { toast } from "sonner";
+  WalletSecurityMethodDialog,
+  WalletSecurityMethodOptions,
+  type WalletSecurityMethod,
+} from "./wallet-security-method-dialog";
 
 interface WalletSecurityClientProps {
   children: ReactNode;
@@ -27,22 +19,8 @@ export function WalletSecurityClient({
   hasVerification,
 }: WalletSecurityClientProps) {
   const [showDialog, setShowDialog] = useState(!hasVerification);
-  const t = useTranslations("private.auth.wallet-security");
-  const router = useRouter();
-
-  const onSubmit = async (data: PincodeFormValues) => {
-    try {
-      setShowDialog(false);
-      await setPincode({
-        name: data.pincodeName,
-        pincode: data.pincode,
-      });
-      toast.success(t("pincode-set"));
-      router.refresh();
-    } catch (error) {
-      console.error("Failed to set pincode:", error);
-    }
-  };
+  const [selectedMethod, setSelectedMethod] =
+    useState<WalletSecurityMethod | null>(null);
 
   return (
     <>
@@ -50,18 +28,22 @@ export function WalletSecurityClient({
         children
       ) : (
         <div className="min-h-screen w-full bg-[url('/backgrounds/background-lm.svg')] bg-center bg-cover dark:bg-[url('/backgrounds/background-dm.svg')]">
-          <Dialog open={showDialog} onOpenChange={setShowDialog}>
-            <DialogContent>
-              <DialogHeader>
-                <DialogTitle>{t("setup-pincode")}</DialogTitle>
-                <DialogDescription>
-                  {t("pincode-instruction")}
-                </DialogDescription>
-              </DialogHeader>
-
-              <PincodeForm onSubmit={onSubmit} />
-            </DialogContent>
-          </Dialog>
+          <WalletSecurityMethodDialog
+            open={showDialog}
+            onOpenChange={() => setSelectedMethod(null)}
+            onSelect={setSelectedMethod}
+          />
+          <PincodeDialog
+            open={selectedMethod === WalletSecurityMethodOptions.Pincode}
+            onOpenChange={setShowDialog}
+          />
+          <SetupTwoFactorDialog
+            onOpenChange={() => setSelectedMethod(null)}
+            open={
+              selectedMethod ===
+              WalletSecurityMethodOptions.TwoFactorAuthentication
+            }
+          />
         </div>
       )}
     </>
